+++ example/lib/page/home/view.dart
@@ -175,6 +175,7 @@ class HomePage extends StatelessWidget {
                           pageBgColor: BaseColors.cTransparent,
                           enableHeroTag: true,
                           singleTop: false,
+                          tag: 't1',
                           bottomView: GestureDetector(
                             onTap: () {
                               Get.find<ImagePreviewLogic>().deleteAt(1);
@@ -193,34 +194,35 @@ class HomePage extends StatelessWidget {
                             ),
                           ),
                         );
-                        // Nav.startImagePreview(
-                        //   imgList: [
-                        //     'https://static1014.gitee.io/pm_data/gallery/images/1/34d40d52a4.jpg',
-                        //     'https://p7.itc.cn/images01/20200529/7fba45f763c445be964badee248ed321.png',
-                        //   ],
-                        //   textTagList: [null, null, (true, null, null)],
-                        //   defaultIndex: 1,
-                        //   singleTop: false,
-                        //   pageBgColor: BaseColors.cTransparent,
-                        //   enableHeroTag: true,
-                        //   bottomView: GestureDetector(
-                        //     onTap: () {
-                        //       Get.find<ImagePreviewLogic>().deleteAt(1);
-                        //     },
-                        //     child: mText(msg: '你好啊'),
-                        //   ),
-                        //   onPreviewIndexChanged: (i) {
-                        //     logic.index(i);
-                        //   },
-                        //   actionView: GestureDetector(
-                        //     onTap: () {
-                        //       logic.list[logic.index.value] = !logic.list[logic.index.value];
-                        //     },
-                        //     child: Obx(
-                        //       () => Icon(logic.list[logic.index.value] ? Icons.check_box : Icons.check_box_outline_blank_outlined),
-                        //     ),
-                        //   ),
-                        // );
+                        Nav.startImagePreview(
+                          imgList: [
+                            'https://static1014.gitee.io/pm_data/gallery/images/1/34d40d52a4.jpg',
+                            'https://p7.itc.cn/images01/20200529/7fba45f763c445be964badee248ed321.png',
+                          ],
+                          textTagList: [null, null, (true, null, null)],
+                          defaultIndex: 1,
+                          singleTop: false,
+                          tag: 't2',
+                          pageBgColor: BaseColors.cTransparent,
+                          enableHeroTag: true,
+                          bottomView: GestureDetector(
+                            onTap: () {
+                              Get.find<ImagePreviewLogic>().deleteAt(1);
+                            },
+                            child: mText(msg: '你好啊'),
+                          ),
+                          onPreviewIndexChanged: (i) {
+                            logic.index(i);
+                          },
+                          actionView: GestureDetector(
+                            onTap: () {
+                              logic.list[logic.index.value] = !logic.list[logic.index.value];
+                            },
+                            child: Obx(
+                              () => Icon(logic.list[logic.index.value] ? Icons.check_box : Icons.check_box_outline_blank_outlined),
+                            ),
+                          ),
+                        );
                       },
                       child: buildImage('https://p7.itc.cn/images01/20200529/7fba45f763c445be964badee248ed321.png', 1),
                     ),